+++ yang/yang-data-api/src/main/java/org/opendaylight/yangtools/yang/data/api/codec/Uint8Codec.java
@@ -8,19 +8,17 @@
 package org.opendaylight.yangtools.yang.data.api.codec;
 
 import org.opendaylight.yangtools.concepts.Codec;
+import org.opendaylight.yangtools.yang.common.Uint8;
 
 /**
- * FIXME: Should be changed to UnsignedByte.
+ * A codec between Uint8 and some other representation.
  *
- * @author ttkacik
- *
- * @param <T>
- *          Output type of serialization.
+ * @param <T> Output type of serialization.
  */
-public interface Uint8Codec<T> extends Codec<T, Short> {
+public interface Uint8Codec<T> extends Codec<T, Uint8> {
     @Override
-    T serialize(Short data);
+    T serialize(Uint8 data);
 
     @Override
-    Short deserialize(T data);
+    Uint8 deserialize(T data);
 }
